attendence_back/requirements.txt
django
gunicorn
pillow
psycopg2-binary
django-leaflet
requests
djangorestframework
